rename: test folder and files
