+++ tests/jsCode/assign.js
