+++ tests/jsCode/fileReader.ts
@@ -1,11 +1,11 @@
 import { readFileSync } from "fs"
 import { fileURLToPath } from "url"
 import { dirname, join } from "path"
-import { convert } from "../tool/converter"
+import { convert } from "../../tool/converter"
 
 export function openJsFile(filename: string): string {
   const __dirname = dirname(fileURLToPath(import.meta.url))
-  const filePath = join(__dirname, "js", filename)
+  const filePath = join(__dirname, filename)
   return readFileSync(filePath, "utf-8")
 }
