+++ tests/jsCode/new.js
